requirements.txt
Flask==3.0.2
flask-paginate==2024.3.28
mysqlclient==2.2.4
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Cython==3.0.10
Flask-WTF==1.2.1
ruff==0.3.4
